+++ pkgs/dart_mcp_server/test_fixtures/dart_cli_app/bin/infinite_wait.dart
@@ -2,9 +2,25 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
+import 'dart:convert';
+import 'dart:developer';
 import 'dart:io';
 
 void main() async {
+  registerExtension('ext.test.echo', (method, parameters) async {
+    return ServiceExtensionResponse.result(
+      jsonEncode({'method': method, 'parameters': parameters}),
+    );
+  });
+
+  registerExtension('ext.test.failure', (method, parameters) async {
+    return ServiceExtensionResponse.error(
+      ServiceExtensionResponse.extensionError,
+      jsonEncode({'message': 'Something went wrong'}),
+    );
+  });
+  print('ready');
+
   stdin.listen((data) {
     if (String.fromCharCodes(data).trim() == 'q') exit(0);
   }, onDone: () => exit(0));